Rename folder `linters` to `linterlib`
